fix: highlight wrapper inside another wrapper not get updated when deleting
When doing multi-color highlight, a highlight(A) created inside another
highlight(B) is not get updated when deleting highlight(A). Though the
highlight id is updated, but class list is not synced with highlight(B)
which makes it to feel like it's not get deleted.

This commit will make it to retain the wrapper classList which associated
with the "extra id".

related issue: #58